refactor(arch): Phase 3 — PR #484 리뷰 3차 수정
- B1: cross-entity import 4곳에 Phase 9 해소 계획 주석 추가
- B2: entities/ticker 순환 의존 해소 (barrel → relative import)
