+++ src/entities/ticker/__tests__/lib/getAssetInfo.test.ts
@@ -49,7 +49,7 @@ jest.mock('@y0ngha/siglens-core', () => ({
 jest.mock('../../lib/db', () => ({
     tryGetTickerDatabaseClient: () => tryGetTickerDatabaseClientMock(),
 }));
-jest.mock('@/entities/ticker', () => ({
+jest.mock('../../api', () => ({
     DrizzleAssetTranslationRepository: class {
         constructor(db: unknown) {
             return repositoryFactoryMock(db) as unknown as object;
